Move Resolver type into separate file
Following comment at https://github.com/urllib3/urllib3/pull/3781/changes#r2867544981, the `Resolver` class was moved to `resolver.py`
