+++ src/urllib3/util/__init__.py
@@ -1,8 +1,9 @@
 # For backwards compatibility, provide imports that used to be here.
 from __future__ import annotations
 
-from .connection import Resolver, is_connection_dropped
+from .connection import is_connection_dropped
 from .request import SKIP_HEADER, SKIPPABLE_HEADERS, make_headers
+from .resolver import Resolver
 from .response import is_fp_closed
 from .retry import Retry
 from .ssl_ import (
